Implement #15 : Folders are now managed
Now even empty folders are managed by the system.
If a patch remove a folder, the update will only remove it if he is not empty.
If a patch remove a file, failing to remove it doesn't fail the whole update.
The copy of a local repository also take care of folders